+++ src/Parser/Visitor/NamespaceCollectingVisitor.php
@@ -1,5 +1,5 @@
 <?php
-namespace J6s\PhpArch\Parser;
+namespace J6s\PhpArch\Parser\Visitor;
 
 
 use PhpParser\NodeVisitorAbstract;