+++ src/Utility/NamespaceComperator.php
@@ -1,6 +1,20 @@
 <?php
 namespace J6s\PhpArch\Utility;
 
+/**
+ * Namespace comperator:
+ * Utility class that can be used in order to compare namespaces.
+ * The comperator has a base namespace that is compared against and can
+ * check whether or not a given other namespace is inside of the base namespace.
+ *
+ * @example
+ * $namespace = new NamespaceComperator('MyVendor\\MyPackage');
+ *
+ * $namespace->contains('MyVendor\\MyPackage\\MyComponent');
+ * // => true
+ * $namespace->contains('MyVendor\\MyOtherPackage\\MyComponent');
+ * // => false
+ */
 class NamespaceComperator
 {
 
@@ -17,11 +31,17 @@ class NamespaceComperator
         return implode('\\', $this->comparison);
     }
 
-    public function matches(string $namespace): bool
+    public function contains(string $namespace): bool
     {
         $parts = explode('\\', trim($namespace, '\\'));
-        $end = min(\count($parts), \count($this->comparison));
 
+        // If the comperator namespace is more specific than the compared namespace
+        // then there is no way that the compared namespace can be inside of it.
+        if (\count($this->comparison) > \count($parts)) {
+            return false;
+        }
+
+        $end = min(\count($parts), \count($this->comparison));
         for ($i = 0; $i < $end; $i++) {
             if ($this->comparison[$i] !== $parts[$i]) {
                 return false;