+++ src/Utility/NamespaceComperatorCollection.php
@@ -2,6 +2,27 @@
 
 namespace J6s\PhpArch\Utility;
 
+/**
+ * Collection of namespace comperators.
+ * Can be used to check whether or not a given namespace is contained
+ * in any of the namespace comperators.
+ *
+ * @see NamespaceComperator
+ * @example
+ * $component = new NamespaceComperatorCollection([
+ *      new NamespaceComperator('App\\Http\\Controllers'),
+ *      new NamespaceComperator('App\\ApiV1\\Controllers')
+ * ]);
+ *
+ * $component->containsAny('App\\Http\\Controllers\\MyController');
+ * // => true
+ *
+ * $component->containsAny('App\\ApiV1\\Controllers\\MyController');
+ * // => true
+ *
+ * $component->containsAny('App\\Utility\\PhoneNumberUtility');
+ * // => false
+ */
 class NamespaceComperatorCollection
 {
 
@@ -24,10 +45,10 @@ class NamespaceComperatorCollection
         );
     }
 
-    public function matchesAny(string $namespace): bool
+    public function containsAny(string $namespace): bool
     {
         foreach ($this->comperators as $comperator) {
-            if ($comperator->matches($namespace)) {
+            if ($comperator->contains($namespace)) {
                 return true;
             }
         }