+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2018-12-01
+### Added
+- Now correctly identifies the following types of dependencies (though in most cases they have already been tracked through `use` statements):
+    - DocBlock comments
+    - Static method calls
+    - Argument type annotations
+    - Return type annotations
+
+### Fixed
+- Namespace comparisons against higher up namespaces no longer match
+
 ## [0.2.0] - 2018-11-30
 ### Added
 - New `assertHasNoErrors` method was added to make checking for architecture violations in PHPUnit easier.